requirements.txt
torch>=2.2.0
torchvision>=0.17.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-image>=0.22.0
matplotlib>=3.8.0
git+https://github.com/openai/CLIP.git 